+++ src/main/java/com/l1yp/service/system/impl/MenuServiceImpl.java
@@ -3,8 +3,11 @@ package com.l1yp.service.system.impl;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.l1yp.exception.VanException;
 import com.l1yp.mapper.system.MenuMapper;
+import com.l1yp.mapper.system.RoleMenuMapper;
 import com.l1yp.model.db.system.Menu;
+import com.l1yp.model.db.system.RoleMenu;
 import com.l1yp.model.db.system.User;
 import com.l1yp.model.param.system.menu.MenuCreateParam;
 import com.l1yp.model.param.system.menu.MenuUpdateParam;
@@ -15,10 +18,14 @@ import com.l1yp.util.RequestUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 @Service
 public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu> implements IMenuService {
+
+    @Resource
+    RoleMenuMapper roleMenuMapper;
 
     @Override
     @Transactional
@@ -47,7 +54,12 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu> implements IM
     }
 
     @Override
+    @Transactional
     public void deleteMenu(Long id) {
+        Long count = roleMenuMapper.selectCount(Wrappers.<RoleMenu>lambdaQuery().eq(RoleMenu::getMenuId, id));
+        if (count > 0) {
+            throw new VanException(500, "请先解绑角色菜单的绑定信息");
+        }
         getBaseMapper().deleteById(id);
     }
